+++ src/Logging/LoggerServiceProvider.php
@@ -7,7 +7,6 @@ namespace Maduser\Argon\Logging;
 use Maduser\Argon\Container\AbstractServiceProvider;
 use Maduser\Argon\Container\ArgonContainer;
 use Maduser\Argon\Container\Exceptions\ContainerException;
-use Monolog\Logger;
 use Psr\Log\LoggerInterface;
 
 /**
@@ -33,15 +32,15 @@ final class LoggerServiceProvider extends AbstractServiceProvider
     {
         $parameters = $container->getParameters();
 
-        if (class_exists($this->loggerClass) && class_exists($this->handlerClass)) {
-            $container->set(LoggerFactory::class, args: [
-                'logLevel' => $parameters->get('logLevel', 200),
-                'logFile' => $parameters->get('logFile', null),
-            ]);
+        $container->set(LoggerFactory::class, args: [
+            'logLevel' => $parameters->get('logLevel', 200),
+            'logFile' => $parameters->get('logFile', null),
+            'loggerClass' => $this->loggerClass,
+            'handlerClass' => $this->handlerClass,
+        ]);
 
-            $container->set(LoggerInterface::class, Logger::class)
-                ->factory(LoggerFactory::class, 'create')
-                ->tag('logger');
-        }
+        $container->set(LoggerInterface::class, LoggerFactory::class)
+            ->factory(LoggerFactory::class, 'create')
+            ->tag('logger');
     }
 }